fix based on comments #3052
