+++ src/Command/Cron/SynchronizeIdossCommand.php
@@ -6,7 +6,7 @@ use App\Entity\JobEvent;
 use App\Entity\Partner;
 use App\Entity\Suivi;
 use App\Entity\User;
-use App\Factory\SuiviFactory;
+use App\Manager\SuiviManager;
 use App\Repository\AffectationRepository;
 use App\Repository\SignalementRepository;
 use App\Repository\UserRepository;
@@ -39,7 +39,7 @@ class SynchronizeIdossCommand extends AbstractCronCommand
         private readonly SignalementRepository $signalementRepository,
         private readonly AffectationRepository $affectationRepository,
         private readonly UserRepository $userRepository,
-        private readonly SuiviFactory $suiviFactory,
+        private readonly SuiviManager $suiviManager,
         private readonly NotificationMailerRegistry $notificationMailerRegistry,
         private readonly ParameterBagInterface $parameterBag,
         private readonly IdossService $idossService,
@@ -132,11 +132,12 @@ class SynchronizeIdossCommand extends AbstractCronCommand
                         default:
                             $description = 'Le signalement a été mis à jour ("'.$item['statut'].'") par IDOSS';
                     }
-                    $suivi = $this->suiviFactory->createInstanceFrom(
+                    $suivi = $this->suiviManager->createSuivi(
                         user: $this->adminUser,
                         signalement: $signalement,
                         description: $description,
                         type: Suivi::TYPE_TECHNICAL,
+                        flush: false
                     );
                     $this->entityManager->persist($suivi);
                 }
