+++ tests/Functional/Manager/SuiviManagerTest.php
@@ -7,37 +7,42 @@ use App\Entity\Signalement;
 use App\Entity\Suivi;
 use App\Entity\User;
 use App\EventListener\SignalementUpdatedListener;
-use App\Factory\SuiviFactory;
 use App\Manager\SuiviManager;
+use App\Repository\DesordreCritereRepository;
 use App\Repository\UserRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Bundle\SecurityBundle\Security;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 class SuiviManagerTest extends KernelTestCase
 {
     private const REF_SIGNALEMENT = '2022-8';
     private ManagerRegistry $managerRegistry;
-    private SuiviFactory $suiviFactory;
     private SignalementUpdatedListener $signalementUpdatedListener;
     private Security $security;
+    private UrlGeneratorInterface $urlGenerator;
+    private DesordreCritereRepository $desordreCritereRepository;
 
     protected function setUp(): void
     {
         self::bootKernel();
         $this->managerRegistry = self::getContainer()->get(ManagerRegistry::class);
-        $this->suiviFactory = static::getContainer()->get(SuiviFactory::class);
         $this->signalementUpdatedListener = static::getContainer()->get(SignalementUpdatedListener::class);
         $this->security = static::getContainer()->get(Security::class);
+        $this->urlGenerator = static::getContainer()->get(UrlGeneratorInterface::class);
+        $this->desordreCritereRepository = static::getContainer()->get(DesordreCritereRepository::class);
     }
 
     public function testCreateSuivi(): void
     {
         $suiviManager = new SuiviManager(
-            $this->suiviFactory,
             $this->managerRegistry,
+            $this->urlGenerator,
             $this->signalementUpdatedListener,
             $this->security,
+            $this->desordreCritereRepository,
             Suivi::class,
         );
 
@@ -59,7 +64,7 @@ class SuiviManagerTest extends KernelTestCase
         $suivi = $suiviManager->createSuivi(
             user : $user,
             signalement : $signalement,
-            description : SuiviFactory::buildDescriptionClotureSignalement($params),
+            description : SuiviManager::buildDescriptionClotureSignalement($params),
             type : Suivi::TYPE_PARTNER,
             isPublic : true,
         );
@@ -68,5 +73,10 @@ class SuiviManagerTest extends KernelTestCase
 
         $this->assertEquals(Suivi::TYPE_PARTNER, $suivi->getType());
         $this->assertNotEquals($countSuivisBeforeCreate, $countSuivisAfterCreate);
+        $this->assertInstanceOf(Suivi::class, $suivi);
+        $desc = 'Le signalement a été cloturé pour test avec le motif suivant <br><strong>Non décence</strong><br><strong>Desc. : </strong>Lorem ipsum suivi sit amet, consectetur adipiscing elit.';
+        $this->assertEquals($desc, $suivi->getDescription());
+        $this->assertTrue($suivi->getIsPublic());
+        $this->assertInstanceOf(UserInterface::class, $suivi->getCreatedBy());
     }
 }
